276 add verified space indicator (#311)
* Add MEW wallet assets
* Add verified DAO flag and mark
* Rework DashboardFollowedDAOsActiveVoteHorizontalListView
* Proposal list item style adjustments
* Update SnapshotProposal status view
* Fix voting power decoding from Int to Double
* Make activeVotes and verified params on Dao not optional
* Fix mark all as read request.
* Get unread count from mark-as-read and mark-as-unread
* Update FollowedDaosDataSource after changing auth token
* Adjust spacing
* Mark notification as read request
* Add Haptic.

+++ Goverland/Goverland/Model/Vote/ProposalAddressValidation.swift
@@ -15,7 +15,7 @@ struct ValidationError: Error, Decodable {
 }
 
 struct ProposalAddressValidation: Decodable {
-    typealias VotingPower = Int
+    typealias VotingPower = Double
 
     let result: Result<VotingPower, ValidationError>
 
@@ -29,7 +29,7 @@ struct ProposalAddressValidation: Decodable {
         let container  = try decoder.container(keyedBy: CodingKeys.self)
 
         let valid = try container.decode(Bool.self, forKey: .ok)
-        let votingPower = try container.decode(Int.self, forKey: .votingPower)
+        let votingPower = try container.decode(Double.self, forKey: .votingPower)
 
         if valid {
             self.result = .success(votingPower)
